add 'real life' spec on chainable matchers
By definition, since it passes, this test only
covers the happy path. But it is slightly more 
integrated than the rest of the tests, which assert
about the results of tests in a different jasmine
environment. It also allows you to observe the 
failing behavior of the chainable matchers, by 
messing with the assertions.