Make chained matchers produce correct expectation results
The ExpectationResult created by the first matcher in
a chain is now passed to all of the successive matchers.
Each matcher updates the result's pass/fail status,
message and stack trace.

Normally, the chain of matchers will produce a passing
result iff *all* of the matchers in the chain match
successfully (i.e. their definition functions return
true).

With a `not` at the beginning of the matcher chain,
the chain will produce a passing result iff *any*
of the matchers in the chain fail to match.