Make getGlobal() work in strict mode
An update of https://github.com/metaweta/jasmine/commit/fb3e1acb09b4dbb9ccdbc15ce64b4a0e471f4541

ES5 strict mode does not promote an undefined 'this' to the global object.  The only way to get the global object in strict mode is to say 'this' while in the global scope.