Insert separators in spec names
When suites are nested, the concatenated spec names can become hard
to follow, e.g. when a failing spec is presented by a reporter.
Including a separator can help understanding the scenario under test
and finding the failing spec more quickly.

For example, change this:

Events when listener is added and added again and removed once is not
notified anymore

to:

Events, when listener is added, and added again, and removed once: is
not notified anymore