Support `jasmine.getEnv().done()` to mark an async test as done outside of the current test lexical scope
There are numerous examples where an external bit of Javascript knows the test has failed and wants the test to fail immediately.  The `jasmine.getEnv().fail()` command is useful, however it does not cause an async test to stop so you have to wait for it to timeout.  This is incredibly inefficient for large test suites.